+++ watercrawl/urls.py
@@ -28,6 +28,6 @@ urlpatterns = [
     path('api/schema/swagger-ui/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
     path('api/schema/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
     path('api/v1/user/', include('user.urls')),
-
+    path('api/v1/common/', include('common.urls')),
     path('api/v1/core/', include('core.urls')),
 ]